move some components to more relevant folders
