+++ next/components/forms/EventReservationForm/index.tsx
@@ -2,7 +2,7 @@ import { EventCardEntityFragment } from '@bratislava/strapi-sdk-city-library'
 import { DateTimeSelect, Input, TextArea } from '@bratislava/ui-city-library'
 import NumberSwitcher from '@bratislava/ui-city-library/NumberSwitcher/NumberSwitcher'
 import { yupResolver } from '@hookform/resolvers/yup'
-import FormatEventDateRange from '@modules/common/FormatEventDateRange'
+import FormatEventDateRange from '@modules/formatting/FormatEventDateRange'
 import { convertDataToBody, getLocalDateForYup } from '@utils/form-constants'
 import { dayForDifferentDateTo, isEventPast } from '@utils/utils'
 import isEmpty from 'lodash/isEmpty'
