+++ next/components/pages/BasicDocumentPage.tsx
@@ -6,7 +6,7 @@ import Home from '@assets/images/home.svg'
 import { BasicDocumentEntity, FooterEntity, MenuEntity } from '@bratislava/strapi-sdk-city-library'
 import { FileIcon, Link, SectionContainer } from '@bratislava/ui-city-library'
 import Button from '@modules/common/Button'
-import FormatDate from '@modules/common/FormatDate'
+import FormatDate from '@modules/formatting/FormatDate'
 import truncate from 'lodash/truncate'
 import { useTranslation } from 'next-i18next'
 import React from 'react'
